backend/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.0.4
SQLAlchemy==2.0.22
requests==2.31.0
scikit-learn==1.5.2
kiwipiepy==0.19.1
pytz==2023.3
openai==0.28.0
ffmpeg==1.4
ffmpeg-python==0.2.0
pydub==0.25.1
PyJWT==2.6.0